+++ console_backend/src/ntrip_output.rs
@@ -1,12 +1,38 @@
 use crate::ntrip_tab::OutputType;
 use crate::types::{ArcBool, Result};
-use crate::utils;
 use anyhow::Context;
 use crossbeam::channel::Receiver;
 use log::error;
 use std::io::Write;
+use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
 use std::{env, io, thread};
+
+pub fn app_dir() -> Result<PathBuf> {
+    let current_exe = std::env::current_exe()?;
+    current_exe
+        .parent()
+        .ok_or(anyhow::format_err!("no parent directory"))
+        .map(Path::to_path_buf)
+}
+
+pub fn pythonhome_dir() -> Result<PathBuf> {
+    let app_dir = app_dir()?;
+    if cfg!(target_os = "macos") {
+        if let Some(parent) = app_dir.parent() {
+            let resources = parent.join("Resources/lib");
+            if resources.exists() {
+                Ok(parent.join("Resources"))
+            } else {
+                Ok(app_dir)
+            }
+        } else {
+            Ok(app_dir)
+        }
+    } else {
+        Ok(app_dir)
+    }
+}
 
 pub struct MessageConverter {
     in_rx: Receiver<Vec<u8>>,
@@ -50,8 +76,8 @@ impl MessageConverter {
 
     /// Runs rtcm3tosbp converter
     fn output_sbp<W: Write + Send + 'static>(&mut self, mut out: W) -> Result<()> {
-        println!("{:?}", utils::app_dir());
-        println!("{:?}", utils::pythonhome_dir());
+        println!("{:?}", app_dir());
+        println!("{:?}", pythonhome_dir());
         println!("{:?}", env::current_exe());
         let mut child = if cfg!(target_os = "windows") {
             let mut cmd = Command::new("cmd");
